Adjust capacity to avoid off by one error
The previous implementation of `capacity` and `capacity_resized` would
floor the results because of integer division truncation. As a result,
when the number was `(1 << N) * 19 / 20 + 1`, the capacity would falsely
be 1 less than expected. For example `Filter::new(61,
0.01).unwrap().capacity()` would actually eqaul `60` instead of `61` as
expected.

This commit looks to fix this issue by first converting the numbers to
floats, then running the `ceil` method, ensuring that the capacity still
works as expected.

Update the test to check up to 32 bits